[Chore] #153 ContactEmailView 구현에 사용된 UIDevice, Bundle extension 로직 파일 분리
