+++ Projects/Coffice/Sources/App/Main/MyPage/Contact/ContactEmailViewState.swift
@@ -8,58 +8,6 @@
 
 import SwiftUI
 
-extension Bundle {
-  func decode<T: Decodable>(
-    _ type: T.Type,
-    from file: String,
-    dateDecodingStategy: JSONDecoder.DateDecodingStrategy = .deferredToDate,
-    keyDecodingStrategy: JSONDecoder.KeyDecodingStrategy = .useDefaultKeys
-  ) -> T {
-    guard let url = CofficeResources.bundle.url(forResource: "DeviceModels", withExtension: "json")
-    else {
-      fatalError("Error: Failed to locate \(file) in bundle.")
-    }
-    guard let data = try? Data(contentsOf: url)
-    else {
-      fatalError("Error: Failed to load \(file) from bundle.")
-    }
-
-    let decoder = JSONDecoder()
-    decoder.dateDecodingStrategy = dateDecodingStategy
-    decoder.keyDecodingStrategy = keyDecodingStrategy
-
-    guard let loaded = try? decoder.decode(T.self, from: data)
-    else {
-      fatalError("Error: Failed to decode \(file) from bundle.")
-    }
-    return loaded
-  }
-}
-
-extension UIDevice {
-  struct DeviceModel: Decodable {
-    let identifier: String
-    let model: String
-    static var all: [DeviceModel] {
-      Bundle.main.decode([DeviceModel].self, from: "DeviceModels.json")
-    }
-  }
-  var modelName: String {
-#if targetEnvironment(simulator)
-    let identifier = ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"]!
-#else
-    var systemInfo = utsname()
-    uname(&systemInfo)
-    let machineMirror = Mirror(reflecting: systemInfo.machine)
-    let identifier = machineMirror.children.reduce("") { identifier, element in
-      guard let value = element.value as? Int8, value != 0 else { return identifier }
-      return identifier + String(UnicodeScalar(UInt8(value)))
-    }
-#endif
-    return DeviceModel.all.first {$0.identifier == identifier }?.model ?? identifier
-  }
-}
-
 struct ContactEmailViewState: Equatable, Identifiable {
   let id = UUID()
   let toAddress: String
